Move to HibernatePoolableElementViewDao
