+++ sqlstore/src/main/java/uk/ac/bbsrc/tgac/miso/persistence/impl/HibernatePoolableElementViewDao.java
@@ -10,14 +10,14 @@ import java.util.function.Consumer;
 import org.hibernate.Criteria;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
-import org.hibernate.criterion.MatchMode;
-import org.hibernate.criterion.Restrictions;
+import org.hibernate.criterion.*;
 import org.hibernate.sql.JoinType;
 import org.hibernate.type.LongType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import uk.ac.bbsrc.tgac.miso.core.data.Workset;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.StorageLocation.LocationUnit;
 import uk.ac.bbsrc.tgac.miso.core.data.impl.view.PoolableElementView;
 import uk.ac.bbsrc.tgac.miso.core.data.type.PlatformType;
@@ -269,4 +269,12 @@ public class HibernatePoolableElementViewDao implements PoolableElementViewDao, 
     criteria.add(Restrictions.eq("tissueType.alias", type));
   }
 
+  @Override
+  public void restrictPaginationByWorksetId(Criteria criteria, long worksetId, Consumer<String> errorHandler) {
+    DetachedCriteria subquery = DetachedCriteria.forClass(Workset.class)
+            .createAlias("libraryAliquots", "libraryaliquot")
+            .add(Restrictions.eq("id", worksetId))
+            .setProjection(Projections.property("libraryaliquot.id"));
+    criteria.add(Property.forName("id").in(subquery));
+  }
 }
